[LIO-Target]: Fix compile warning on powerpc and x86_64
This patch fixes the following compile warning on powerpc and x86_64:

drivers/lio-core/iscsi_target_configfs.c: In function ‘lio_target_call_addnptotpg’:
drivers/lio-core/iscsi_target_configfs.c:273: warning: format ‘%d’ expects type ‘int’, but argument 2 has type ‘__kernel_size_t’

+++ drivers/lio-core/iscsi_target_configfs.c
@@ -270,7 +270,7 @@ static struct config_group *lio_target_call_addnptotpg(
 	}
 	if (strlen(name) > MAX_PORTAL_LEN) {
 		printk(KERN_ERR "strlen(name): %d exceeds MAX_PORTAL_LEN: %d\n",
-			strlen(name), MAX_PORTAL_LEN);
+			(int)strlen(name), MAX_PORTAL_LEN);
 		return ERR_PTR(-EOVERFLOW);
 	}
 	memset(buf, 0, MAX_PORTAL_LEN);
